+++ debug.py
@@ -6,5 +6,5 @@ from example_view import ExampleView
 
 
 if __name__ == "__main__":
-        game = ExampleGame(ExampleView)
-        game.run()
+    game = ExampleGame(ExampleView)
+    game.run()