+++ pygametemplate/__init__.py
@@ -2,7 +2,7 @@
 from __future__ import absolute_import
 
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 
 __url__ = "https://github.com/AndyDeany/pygame-template"
 __download_url__ = "{}/archive/v{}.tar.gz".format(__url__, __version__)
@@ -11,5 +11,6 @@ __author__ = "Andrew Dean"
 __author_email__ = "oneandydean@hotmail.com"
 
 
+from pygametemplate.core import log, load_image, load_font
 from pygametemplate.game import Game
 from pygametemplate.view import View