+++ pygametemplate/game.py
@@ -2,10 +2,6 @@ from __future__ import absolute_import
 
 
 import os
-import sys
-import ctypes
-import traceback
-import datetime
 import time
 
 import pygame
@@ -16,9 +12,8 @@ try:
 except ImportError:
     pass
 
-from pygametemplate.exceptions import CaughtFatalException
+from pygametemplate import log, load_image
 from pygametemplate.system import System
-from pygametemplate.helper import Helper
 from pygametemplate.console import Console
 from pygametemplate.userinput import Input
 from pygametemplate.hotkey import Hotkey
@@ -27,19 +22,17 @@ from pygametemplate.text_input import TextInput
 
 class Game(object):
     def __init__(self, StartingView, resolution=(1280, 720), mode="windowed"):
-        self.directory = os.getcwd()
-
         try:
             pygame.init()
             self.pygame = pygame
         except Exception:
-            self.log("Failed to initialise pygame")
+            log("Failed to initialise pygame")
         self.system = System(self)
         self.width, self.height = resolution
         self.mode = mode
         self.initialise_screen()
         pygame.display.set_caption("insertnamehere (Alpha 1.0)")
-        #! pygame.display.set_icon(self.load_image("icon_name", file_extension=".ico"))
+        #! pygame.display.set_icon(load_image("icon_name", file_extension=".ico"))
 
         self.last_view = None
         self.current_view = StartingView(self)
@@ -48,7 +41,6 @@ class Game(object):
         self.frame = 0  # The current frame the game is on (since the game was opened)
 
         self.input = Input(self)
-        self.helper = Helper(self)
         self.console = Console(self)
 
         self.quit_condition = Hotkey(self, "f4", alt=True).pressed
@@ -84,32 +76,6 @@ class Game(object):
         self.quit()
         pygame.quit()
 
-    def path_to(self, *path):
-        """Returns the complete absolute path of the path given."""
-        return os.path.join(self.directory, *"/".join(path).split("/"))
-
-    def log(self, *error_message, **options):
-        """Takes 1 or more variables and concatenates them to create the error message."""
-        fatal = options.get("fatal", True)  # `fatal` option defaults to True
-        error_message = "".join(map(str, error_message))
-        try:
-            with open(self.path_to("log.txt"), "a") as error_log:
-                error_log.write("%s - %s.\n" % (datetime.datetime.utcnow(), error_message))
-                error_log.write(traceback.format_exc() + "\n")
-        except Exception:
-            error_info = "This error occurred very early during game initialisation and could not be logged"
-        else:
-            error_info = "Please check log.txt for details"
-
-        if fatal:
-            text = "".join(("An error has occurred:\n\n    ",
-                            error_message, ".\n\n\n",
-                            error_info, "."))
-            ctypes.windll.user32.MessageBoxA(0, text, "Error", 0)   # Error popup
-            raise CaughtFatalException(sys.exc_info()[1])
-        else:
-            pass    #! Add some code here to show an error message in game
-
     def initialise_screen(self, resolution=None, mode=None):
         """(Re)initialises the screen using the given arguments."""
         try:
@@ -132,33 +98,8 @@ class Game(object):
             self.width, self.height = resolution
             self.mode = mode
         except Exception:
-            self.log("Failed to reinitialise screen in ", mode, " mode "
+            log("Failed to reinitialise screen in ", mode, " mode "
                      "at ", self.width, "x", self.height, " resolution")
-
-    # Asset loading
-    def load_image(self, image_name, fade_enabled=False, file_extension=".png"):
-        """fade_enabled should be True if you want images to be able to fade"""
-        try:
-            #! Add stuff for loading images of the correct resolution
-            # depending on the player's resolution settings
-            if not fade_enabled:
-                return pygame.image.load(
-                    self.path_to("assets/images", image_name + file_extension)
-                    ).convert_alpha()   # Fixes per pixel alphas permanently
-            else:
-                return pygame.image.load(
-                    self.path_to("assets/images", image_name + file_extension)
-                    ).convert()
-        except Exception:
-            self.log("Failed to load image: ", image_name, file_extension)
-
-    def load_font(self, font_name, font_size, file_extension=".ttf"):
-        try:
-            return pygame.font.Font(
-                self.path_to("assets/fonts", font_name + file_extension), font_size
-                )
-        except Exception:
-            self.log("Failed to load font: ", font_name, file_extension)
 
     def display(self, image, coordinates, area=None, special_flags=0):
         """Takes coordinates and area for a 1920x1080 window"""
@@ -171,7 +112,7 @@ class Game(object):
                         area[2]*x_scale, area[3]*y_scale)
             self.screen.blit(image, coordinates, area, special_flags)
         except Exception:
-            self.log("Failed to display image at ", coordinates)
+            log("Failed to display image at ", coordinates)
 
     def _inputs(self):
         self.input.reset()
@@ -197,13 +138,13 @@ class Game(object):
             pygame.display.flip()   # Updating the screen
             self.clock.tick(self.fps)    # [fps] times per second
         except Exception:
-            self.log("Failed to update screen")
+            log("Failed to update screen")
 
     def runtime(self):
         try:
             return time.time() - self.start_time
         except Exception:
-            self.log("Failed to calculate and return game run time")
+            log("Failed to calculate and return game run time")
 
     def _check_quit(self):
         if self.quit_condition():
@@ -215,7 +156,7 @@ class Game(object):
             self.clock = pygame.time.Clock()
             self.start_time = time.time()
         except Exception:
-            self.log("Failed to initialise essential time related display variables")
+            log("Failed to initialise essential time related display variables")
 
         while self.running:
             self._inputs()