+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 0.4.0
+
+* [**Backwards incompatible change**]<br>
+Move the `path_to()`, `log()`, `load_image()` and `load_font()` methods of
+the `Game` class to be functions of a new `pygametemplate.core` module.
+`log()`, `load_image()` and `load_font()` are also accessible as direct
+members of the `pygametemplate` module.
+
+* [**Backwards incompatible change**]<br>
+Remove `pygametemplate.helper.Helper` class and move its methods to
+be functions as members of the `pygametemplate.helper` module.
+`Game` instances no longer have a `helper` attribute, meaning that all
+references to `game.helper.method()` will have to be replaced with
+`pygametemplate.helper.method()`.
+
+
 ## 0.3.0
 
 * [**Backwards incompatible change**]<br>
@@ -9,7 +25,6 @@ Every view should inherit from `pygametemplate.View`.
 All `View`s have a `self.game` member for accessing the main `Game` instance.
 To switch to the next view in your game, you can call `self.game.set_view(NextViewClass)`.
 `View` classes are expected to implement the following methods:
-
     * `View.load(self)`: load all the assets/other things needed for the view to work.
     This means you only load images (for example) into RAM when you need them.
 
@@ -21,7 +36,6 @@ To switch to the next view in your game, you can call `self.game.set_view(NextVi
 
     * `View.draw(self)`: run all drawing code for the view.
 
-
 * Change `_inputs()`, `_update()` and `_check_quit()` to be "private" methods
 of the `Game` class by prepending an underscore to their names.
 
@@ -29,6 +43,7 @@ of the `Game` class by prepending an underscore to their names.
 (defaults to `ctrl+f`) so that it can be customised.
 
 * Add 1280x720 as the default window resolution if one isn't passed in.
+
 
 ## 0.2.0